ENAS fixed model can now be trained on TPUs with Average Weight-drop SGD (AWD-SGD; https://arxiv.org/pdf/1708.02182.pdf).
The implementation of ASGD is a bit different. Instead of starting ASGD after
validation not improving, in this implementation, ASGD is scheduled to start at
a pre-defined step, which has to be set as a hyper-parameter. This change still
delivers strong performance (`56.1` perplexity on Penn Treebank), and is more
TPU-friendly.

PiperOrigin-RevId: 245977862