Add an apply_mask method that uses the params passed to the Pruning object.
Currently apply_mask() is a separate method, but uses one of the params ('prune_option') that the Pruning object stores. This makes it inconvenient to invoke apply_mask through an API, making it necessary to store the prune_option separately.

Adding a method to the class to ensure that like other methods related to pruning, apply_mask can also be invoked without the need to store params separately.

PiperOrigin-RevId: 294253907