Adds tests for lexer errors and invalid character errors.
This CL also separates the extraction of tokens from javalang (i.e., the consumption of the generator of `JavaToken`s) from their processing. That way if an exception occurs, the whole sequence is reported as erroneous.

The only error captured and represented in an otherwise correct token sequence is that of an invalid character, which is emitted amidst the other tokens.

A future change will emit more specific tokens for other errors (i.e., LexerError exceptions).

PiperOrigin-RevId: 337048670