Fix nondeterminism in identifier choice during GFSA python code generation.
The original implementation is nondeterministic between processes, because it
relies on the iteration order for Python sets and thus on the seed for the
Python hash function. This change instead fixes a particular order for the
identifiers.
(see https://docs.python.org/3/reference/datamodel.html#object.__hash__)

Also adds some documentation regarding the python code generation used in the
paper.

PiperOrigin-RevId: 343857550