Changes to KWS streaming framework to support U-Net models that are not fully causal:
delay.py:
Add an option to apply delay in non-streaming mode as well as streaming.

temporal_padding.py:
Add support for 'future' padding (i.e., on the RHS) and cropping when padding_size is negative.

stream.py:
Minor update to ring buffer size, so that we never drop/ignore the most recent (i.e. right-most) `stride - 1` samples in case of strided convolution, assuming the input length is a multiple of the stride (which should always be the case in streaming).

conv1d_transpose.py:
Add the option for 'same' padding (non-causal, has lookahead), since `padding` for the Keras super class always needs to be 'valid' in order for self.overlap to be correct. (No longer a necesity with the current approach for lookahead, but could still be useful to have in the future.)

test_utils.py:
Minor: Make sure the result is an int even if cnn_strides is empty.
PiperOrigin-RevId: 346574361