Update gradient for floating-point downcast.
This refactors the custom JVP operation to be written in terms of the Jax-calculted gradient of a (crude) differentiable approximation of the downcasting operation, instead of expicitly implementing the gradient calculations. This is less error-prone and avoids the reinventing the wheel of Jax autodiff.

It also switches to using the STE for all rounding operatons in the downcast, including the rounding of numbers below the lowest FP value to zero. Before, the gradient was hard-coded as zero for those values below the flush-to-zero bound, but on second thought there is no fundamental reason to treat that rounding as different than any other rounding in the downcasting operations with respect to gradient calculation and so the STE is used there as well now.

Experimentally, this leads to an average gain of 1.9 BLEU points in experiments on machine translation with layernorm and softmax quantized.

PiperOrigin-RevId: 362087448