Migrate users of jspb.map.Map.prototype.getEntryList to alternate map apis.
This method is being deleted as it is incompatible with the es6 map api and has some odd corner cases.

Most users are being directly migrate to the entries method or implicit iteration of the map itself. In other cases I migrated directly to a more suitable map API (e.g. keys,foreach,...).

PiperOrigin-RevId: 412940930