Make `swsft_forward_spins_channels` and `_compute_Gnm` faster.
By avoiding vmap and for loops, we avoid repeated computations on
constants and slicing and concatenating over spins in
`swsft_forward_spins_channels`.

We also stop making use of symmetry to reduce the number of operations
in `_compute_Gnm` by default, which performs better on TPU (similar to
what was done in the previous commit).

These changes speed-up training of a large model on TPU by >10%.

PiperOrigin-RevId: 459088506