Fix padding error for matrix_inverse_pth_root.
Because statistics are padded to the block size by the identity matrix in the bottom-right corner, the effective maximum eigenvalue of matrices handed to the pth-root inversion routine is 1.0. This can be much larger than the actual maximum eigenvalue, which can be empirically observed to be much smaller, like 1e-6, in LLM use cases for some parameters.

This can result in counterintuitive behavior where the principal minor of a padded matrix doesn't have the same inverse p-th root as the unpadded version.

This CL adds a corresponding test and ensures this behavior by explicitly zeroing out padding columns & rows for inputs.

PiperOrigin-RevId: 467311420