bugfix (rl_metrics_aaai2021): corrected discrete metric computations.
Updated the DiscreteBisimulation and DiscreteLaxBisimulation metrics to be consistent with their definitions.
In the original implementation the condition check on the transition function was incorrect, as the check was summing transitions over all states in S, instead of checking sums over each equivalence classes.

To minimise code duplication, introduced the function _update_classes(), which should be shared among all discrete metrics.(metrics)