jspb: write type coercions on maps.
This change adds type assertions and coercions to map keys and values, to ensure that they can be successfully parsed by servers and serialized to binary.

As a result of this change, a small number of files with clearly incorrect key or value types needed to be adjusted. Given their size and the correctness of a conversion, this change also fixes them inline.

PiperOrigin-RevId: 518652795