Correct implementation of estimate_topic_top_k_prob.
This change fixes two mistakes in the estimate_topic_top_k_prob function:

1. The expression for \hat{p}[o] in our paper is (1/n)*sum_{i=1}^n (1{W_i^s = o) - q_in) / (q_in - q_out), but the expression implemented in the code was (1/n) * sum_{i=1}^n (1{W_i^s = o) * q_in) / (q_in - q_out).

2. There was a typo in the derivation of the expression in the paper, and the q_in in the numerator of the summands should actually be q_out.

We additionally implement a statistical test to verify the correctness of estimate_topic_top_k_prob. This also gives empirical support for the corrected version of Lemma 10 from the paper.

We verified that this fix does not affect the overall trend of the results from our paper significantly.

PiperOrigin-RevId: 551522301