Fixed error in computation of M
tldr: slight error in computation of support of truncated normal, only effects result for very big or very specific values of theta. 

As stated in the supplementary of the paper by Theis & Yosri,
to compute M we need to compute the difference between 
c = Phi_{sigma^2 + 1}( Phi^{-1}( theta'/2 )) and
d = Phi_{sigma^2 + 1}( Phi^{-1}( 1 - theta'/2 )).

The correct value is thus computed e.g. via
`c = p.cdf(sp.stats.norm().ppf(D / 2.0))`.
The value of c in this notebook is slightly different due to a bug in the computation of the correct q_tr.a.

q.std(), the st.dev. of the target distribution q (here: 1), is slightly larger than q_tr.std(),  the st.dev. of the truncated target distribution (here: ~0.9996), because of the missing distribution tails. To compute the correct support we need to use the former. To check the correctness of q_tr.a one can verify that for a small value eps (e.g. 10^{-15}) q_tr.pdf(q_tr.a - eps) is zero while q_tr.pdf(q_tr.a + eps) is nonzero. This is not true for the current q_tr.a but is true for the proposed fixed version.

Note that due to the flooring operation this does not change M for small theta like `D=1e-4`, but makes a difference for big theta e.g. `D=5e-3` and some specific theta, e.g. `D=2.82e-5`.