Fixed default parameter for target_slice in RevIn layer.
This change fixes an issue with the shape of the target_slice in the call function of the RevIn layer. The default target_slice=None produces the wrong output shape if no target slice is provided. target_slice=(1,None) in combination with slight changes in the shapes of the affine_weight and the affine_bias solves this issue. 

See https://github.com/google-research/google-research/issues/1812#issue-1949990031 for more details.