[Anthea] Fix the result parsing problem caused by document index
In the SxS MQM template, results from both translations of the same source are currently saved with the same document index, resulting in a parsing error in Marot. The error makes all results being attributed to translation 1 and no results to translation 2. This CL fixes this issue by increasing the doc idx of translation 1 by *2 and the doc idx of translation 2 by *2 + 1.

PiperOrigin-RevId: 654151735