[Anthea] Fix the result parsing problem caused by omission errors
In the current SxS MQM implementation, when saving results, annotations on translation and translation2 are split and annotations on source are attributed to both translations. This is problematic for omission errors because it is not always the case that both translations omitted the same content in the source. This CL fixes this issue by having two types of omission errors and adding a property "which_translation_side" to them in template-mqm-sxs.js. The property is used in anthea_eval.js to decide the splitting and merging of errors. Other errors marked on the source side are not affected.

PiperOrigin-RevId: 658519554