[Marot] Support filtering category names with parentheses.
Characters that are significant in regular expressions (most notably parentheses) would previously present problems if present in e.g. an error category name: to filter by them, those characters would need to be escaped. This affected both clicking on an element of a result row, and selecting a filter value from the drop-down menus.

Now, special characters are escaped in both cases.

PiperOrigin-RevId: 676078843