[Anthea] Add better support for bidirectional text.
Bidirectional text is sometimes improperly supported. Specifically, when right-to-left text has embedded left-to-right text at the beginning, the browser assumes that the prevailing order is left-to-right. This change uses the source- and target-language information to infer whether the prevailing direction should be right-to-left.

The list of right-to-left languages and their codes that are supported with this change is:
 - Arabic (ar)
 - Persian (fa)
 - Hebrew (he, iw)
 - Urdu (ur)
 - Pashto (ps)
 - Yiddish (yi)

Note that, with this change, if either the source or target is labeled as a right-to-left language but the text is actually entirely or mostly left-to-right, it will be displayed incorrectly. Users should make sure that language labels are accurate.

PiperOrigin-RevId: 694580275