Fix flaky `symbolic/instructions_test.py`.
The test was flaky because the `assertAlmostEqual` function from `unittest` module doesn't work correctly with `sympy` floats. Inside the `unittest` module, the difference between values is calculated, rounded and ultimately compared to zero-valued `int`. That comparison always returns `False` for `sympy` float, even when its value is `0.0`.

Due to the above, when at least one of the arguments of `assertAlmostEqual` is `sympy` float, the function raises an exception in case of any difference between the values, ignoring the requested tolerance of 7 decimal places.

The issue seems like a bug in the `sympy` library. As a workaround, we can convert `sympy` float to built-in python float, it solves the issue.

PiperOrigin-RevId: 700735281